Split mfa and security components
