--- jsapp/js/account/security/fun.module.scss
@@ -1,8 +1,0 @@
-
-.aqua { 
- background-color: aqua !important;
-}
-
-h2.aqua {
-    background-color: purple;
-}
